+++ src/components/Order/Coupon.tsx
@@ -3,14 +3,8 @@ import { FlexColumn, FlexRow } from "@/style/common.style";
 import { COUPON_MESSAGE } from "@/constants/message";
 import CheckBox from "../common/CheckBox";
 import { CouponType } from "@/types/coupon.type";
-import { MAX_APPLICABLE_COUPON } from "@/constants/system";
-import { cartSummaryState } from "@/store/selectors/summarySelector/cartSummarySelector";
-import { curKoreaTime } from "@/utils/date";
-import { isValidExpirationDate } from "@/utils/isValidExpirationDate";
-import { selectedItemsState } from "@/store/selectors/selectedSelector/selectedItemsSelector";
 import styled from "@emotion/styled";
-import { useRecoilValue } from "recoil";
-import { useState } from "react";
+import useCoupon from "@/hooks/useCoupon";
 
 interface Props {
   coupon: CouponType;
@@ -19,64 +13,11 @@ interface Props {
 }
 
 const Coupon = ({ coupon, selectedCoupons, setSelectedCoupons }: Props) => {
-  const [isSelected, setSelected] = useState(selectedCoupons.includes(coupon));
-  const { orderPrice } = useRecoilValue(cartSummaryState);
-  const selectedItems = useRecoilValue(selectedItemsState);
-
-  const handleClick = () => {
-    setSelected(!isSelected);
-
-    if (!isSelected && !selectedCoupons.includes(coupon)) {
-      const newList = [...selectedCoupons, coupon];
-      setSelectedCoupons(newList);
-      return;
-    }
-
-    const filteredList = selectedCoupons.filter(
-      (selectedCoupon) => coupon.id !== selectedCoupon.id
-    );
-
-    setSelectedCoupons(filteredList);
-  };
-
-  const checkDisabled = () => {
-    //유효기간 확인
-    if (!isValidExpirationDate(coupon.expirationDate)) {
-      return true;
-    }
-    // 사용 가능 시간 확인
-    if (coupon.availableTime) {
-      const startTime = Number(coupon.availableTime.start.slice(0, 2));
-      const endTime = Number(coupon.availableTime.end.slice(0, 2));
-      const curTime = curKoreaTime.getHours();
-
-      if (curTime < startTime || curTime >= endTime) {
-        return true;
-      }
-    }
-
-    //최소 주문 금액 확인
-    if (coupon.minimumAmount && orderPrice < coupon.minimumAmount) {
-      return true;
-    }
-
-    // BOGO 쿠폰 최소 수량 확인
-    if (selectedItems.every((item) => item.quantity < 3)) {
-      return true;
-    }
-
-    // coupon 최대 적용 갯수 확인
-    if (
-      selectedCoupons.length >= MAX_APPLICABLE_COUPON &&
-      !selectedCoupons.includes(coupon)
-    ) {
-      return true;
-    }
-
-    return false;
-  };
-
-  const disabled = checkDisabled();
+  const { isSelected, disabled, handleSelect } = useCoupon({
+    coupon,
+    selectedCoupons,
+    setSelectedCoupons,
+  });
 
   return (
     <StyledItemWrapper disabled={disabled}>
@@ -84,7 +25,7 @@ const Coupon = ({ coupon, selectedCoupons, setSelectedCoupons }: Props) => {
         <CheckBox
           id={`coupon-${coupon.id}`}
           isSelected={isSelected}
-          onClick={disabled ? () => {} : handleClick}
+          onClick={disabled ? () => {} : handleSelect}
           disabled={disabled}
         />
         <StyledBoldText>{coupon.description}</StyledBoldText>
